+++ tests/apps/courses/test_templates_category_detail.py
@@ -248,11 +248,11 @@ class CategoryCMSTestCase(CMSTestCase):
         # Make sure we do have 3 courses on the category
         self.assertEqual(category.get_courses().count(), 3)
 
-        # Only the first two are rendered in the template
+        # Only the last two are rendered in the template
         response = self.client.get(category.extended_object.get_absolute_url())
-        self.assertContains(response, courses[0].extended_object.get_title())
+        self.assertNotContains(response, courses[0].extended_object.get_title())
         self.assertContains(response, courses[1].extended_object.get_title())
-        self.assertNotContains(response, courses[2].extended_object.get_title())
+        self.assertContains(response, courses[2].extended_object.get_title())
 
         # There is a link to view more related courses directly in the Search view
         self.assertContains(
@@ -304,11 +304,11 @@ class CategoryCMSTestCase(CMSTestCase):
         self.assertEqual(category.get_courses().count(), 3)
 
         # NB: here we are requesting the meta category's page, not the child category
-        # Only the first two are rendered in the template
+        # Only the last two are rendered in the template
         response = self.client.get(meta.extended_object.get_absolute_url())
-        self.assertContains(response, courses[0].extended_object.get_title())
+        self.assertNotContains(response, courses[0].extended_object.get_title())
         self.assertContains(response, courses[1].extended_object.get_title())
-        self.assertNotContains(response, courses[2].extended_object.get_title())
+        self.assertContains(response, courses[2].extended_object.get_title())
 
         # There is a link to view all courses
         self.assertContains(response, 'href="/the/courses/')